requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.8.1
atpublic==3.0.1
attrs==21.4.0
blis==0.7.6
boto3==1.21.19
botocore==1.24.19
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==36.0.1
cymem==2.0.6
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
dpath==2.0.6
dulwich==0.20.34
dvc==2.9.4
dvc-render==0.0.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
flatten-dict==0.4.2
flufl.lock==7.0
frozenlist==1.3.0
fsspec==2022.2.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
grandalf==0.6
idna==3.3
importlib-resources==5.4.0
jellyfish==0.9.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
langcodes==3.3.0
langdetect==1.0.9
mailchecker==4.1.14
MarkupSafe==2.1.1
multidict==6.0.2
murmurhash==1.0.6
nanotime==0.5.2
networkx==2.7.1
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathspec==0.9.0
pathy==0.6.1
phonenumbers==8.12.45
ply==3.11
preshed==3.0.6
psutil==5.9.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pydot==1.4.2
pygit2==1.9.0
Pygments==2.11.2
pygtrie==2.4.2
pyparsing==3.0.7
python-benedict==0.25.0
python-dateutil==2.8.2
python-fsutil==0.6.0
python-slugify==6.1.1
pytz==2021.3
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
rich==12.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
scmrepo==0.0.7
shortuuid==1.0.8
shtab==1.5.3
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
-e git+https://github.com/dk6304273089/sentiment-analysis.git@aecce9876f23f1ed6962b067cd4743655d6a17c8#egg=src
srsly==2.4.2
tabulate==0.8.9
text-unidecode==1.3
textblob==0.17.1
thinc==8.0.15
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.63.1
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
vaderSentiment==3.3.2
voluptuous==0.12.2
wasabi==0.9.0
wcwidth==0.2.5
wincertstore==0.2
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.7.0
